requirements.txt
openai-whisper
fastapi
uvicorn[standard]
ffmpeg-python
torch==2.1.0+cu118
-f https://download.pytorch.org/whl/torch_stable.html
